added jwt claims and additional authorization

+++ resources/item.py
@@ -1,7 +1,8 @@
+from email import message
 from flask import request
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required,get_jwt
 from sqlalchemy.exc import SQLAlchemyError
 from db import db
 from models.item import ItemModel
@@ -19,6 +20,10 @@ class Item(MethodView):
 
     @jwt_required()
     def delete(self, item_id):
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilage required.")
+
         item = ItemModel.query.get_or_404(item_id)
         db.session.delete(item)
         db.session.commit()
